+++ backend/src/algorithm/index.ts
@@ -1,14 +1,7 @@
-import {
-  checkConstraints,
-  getClassPeriodsPerDay,
-  getMatrixHashmap,
-  getTimesPerWeek,
-  setSlot,
-  shuffleArray,
-  swapSlots,
-} from './utils';
+import { checkConstraints, getClassPeriodsPerDay, getTimesPerWeek, setSlot, shuffleArray, swapSlots } from './utils';
 import { daysPerWeek, maxPeriodsPerDay } from './consts';
 import { GenerateTimetableProps, GenerateTimetableResult, RemainingLectures } from './types';
+import { getMatrixHashmap } from './utils/matrix';
 
 export function generateTimetable({ classes, classrooms }: GenerateTimetableProps): GenerateTimetableResult {
   const teachers = classes
