+++ backend/src/algorithm/utils/matrix.ts
@@ -1,0 +1,16 @@
+import { daysPerWeek, maxPeriodsPerDay } from 'algorithm/consts';
+import { hash } from './hash';
+
+export function getMatrix(x: number, y: number) {
+  const matrix = new Array(x);
+
+  for (let i = 0; i < x; ++i) {
+    matrix[i] = new Array(y).fill(null);
+  }
+
+  return matrix;
+}
+
+export function getMatrixHashmap(keys: unknown[]) {
+  return Object.fromEntries(keys.map(key => [hash(key), getMatrix(daysPerWeek, maxPeriodsPerDay)]));
+}
